--- src/aeos/owl_aeos_tuner_cos_stub.c
@@ -1,45 +1,0 @@
-/*
- * OWL - OCaml Scientific and Engineering Computing
- * Copyright (c) 2016-2018 Liang Wang <liang.wang@cl.cam.ac.uk>
- */
-
-int dummy_136 = 0;
-
-#include <owl_macros.h>
-
-#define FUN4 float32_cos
-#define NUMBER float
-#define NUMBER1 float
-#define MAPFN(X) (cosf(X))
-
-CAMLprim value FUN4(value vN, value vX, value vY) {
-  CAMLparam3(vN, vX, vY);
-  int N = Long_val(vN);
-
-  struct caml_ba_array *X = Caml_ba_array_val(vX);
-  NUMBER *X_data = (NUMBER *) X->data;
-
-  struct caml_ba_array *Y = Caml_ba_array_val(vY);
-  NUMBER1 *Y_data = (NUMBER1 *) Y->data;
-
-  NUMBER *start_x, *stop_x;
-  NUMBER1 *start_y;
-
-  caml_release_runtime_system();  /* Allow other threads */
-
-  start_x = X_data;
-  stop_x = start_x + N;
-  start_y = Y_data;
-
-  while (start_x != stop_x) {
-    NUMBER x = *start_x;
-    *start_y = MAPFN(x);
-
-    start_x += 1;
-    start_y += 1;
-  };
-
-  caml_acquire_runtime_system();  /* Disallow other threads */
-
-  CAMLreturn(Val_unit);
-}
