--- src/aeos/owl_aeos_tuner_map_sig.ml
@@ -1,41 +1,0 @@
-(* module type Sig = sig
-
-module Sin : sig
-
-  type t = {
-    mutable c_macro : string;
-    mutable params  : int;
-    mutable x       : int array;
-  }
-
-  val make : unit -> t
-
-  val tune : t -> unit
-
-  val to_string : t -> string
-
-end
-
-module Cos : sig
-
-  type t = {
-    mutable c_macro : string;
-    mutable params  : int;
-    mutable x       : int array;
-  }
-
-  val make : unit -> t
-
-  val tune : t -> unit
-
-  val to_string : t -> string
-
-end
-
-type tuner =
-  | Sin of Sin.t
-  | Cos of Cos.t
-
-end
-
-*)
