--- src/aeos/owl_aeos_tuner_sin_sig.ml
@@ -1,22 +1,0 @@
-module type Sig = sig
-
-module Sin : sig
-
-  type t = {
-    mutable c_macro : string;
-    mutable params  : int;
-    mutable x       : int array;
-  }
-
-  val make : unit -> t
-
-  val tune : t -> unit
-
-  val to_string : t -> string
-
-end
-
-type tuner =
-  | Sin of Sin.t
-
-end
