+++ src/owl/config/configure.ml
@@ -82,7 +82,7 @@ let get_gcc_path _c =
 
 
 let get_openblas_path _c =
-  let p0 = "/usr/local/opt/openblas/lib" in
+  let p0 = "/usr/local/lib/openblas/lib" in
   let p1 = "/opt/OpenBLAS/lib/" in
   if Sys.file_exists p0 then ["-L" ^ p0]
   else if Sys.file_exists p1 then ["-L" ^ p1]
